move export logic from build.py to export.py
