+++ .github/workflows/template_build.yml
@@ -110,7 +110,7 @@ jobs:
       - name: Export
         run: |
           # build tables
-          python -m dcpy lifecycle builds build export --recipe-path ${{ inputs.recipe_file }}.lock.yml
+          python -m dcpy lifecycle builds export export --recipe-path ${{ inputs.recipe_file }}.lock.yml
           # metadata files
           python -m build_scripts.export
