+++ dcpy/lifecycle/builds/_cli.py
@@ -2,11 +2,13 @@ import typer
 
 from dcpy.lifecycle.builds.artifacts._cli import app as artifacts_app
 from dcpy.lifecycle.builds.build import app as build_app
+from dcpy.lifecycle.builds.export import app as export_app
 from dcpy.lifecycle.builds.load import app as load_app
 from dcpy.lifecycle.builds.plan import app as plan_app
 
 app = typer.Typer()
 app.add_typer(plan_app, name="plan")
-app.add_typer(build_app, name="build")
 app.add_typer(load_app, name="load")
+app.add_typer(export_app, name="export")
+app.add_typer(build_app, name="build")
 app.add_typer(artifacts_app, name="artifacts")
